refactor: API 중심의 도메인 패키지 구조 개편 및 서비스 레이어 분리
