+++ apps/server/src/main/java/com/weanet/server/api/route/dto/response/PoiResponse.java
@@ -1,0 +1,17 @@
+package com.weanet.server.api.route.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PoiResponse {
+    private String name;
+    private String address;
+    private double lat;
+    private double lng;
+}
