+++ apps/server/src/main/java/com/weanet/server/api/route/dto/response/SubwayRealtimeResponse.java
@@ -1,0 +1,16 @@
+package com.weanet.server.api.route.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class SubwayRealtimeResponse {
+    private String lineName;
+    private String arrivalMessage;
+    private String status;
+}
