+++ apps/server/src/main/java/com/weanet/server/api/route/service/RouteService.java
@@ -5,7 +5,7 @@ import com.weanet.server.api.route.domain.Route;
 import com.weanet.server.api.route.domain.RouteStep;
 import com.weanet.server.api.route.dto.request.RouteSaveRequest;
 import com.weanet.server.api.route.dto.response.*;
-import com.weanet.server.dto.RouteSearchRequest;
+import com.weanet.server.api.route.dto.request.RouteSearchRequest;
 import com.weanet.server.api.common.exception.BusinessException;
 import com.weanet.server.api.common.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
