+++ apps/server/src/main/java/com/weanet/server/api/route/service/SubwayService.java
@@ -1,6 +1,6 @@
 package com.weanet.server.api.route.service;
 
-import com.weanet.server.dto.SubwayRealtimeResponse;
+import com.weanet.server.api.route.dto.response.SubwayRealtimeResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
